Backend/requirements.txt
flask
scikit-learn
pandas
numpy
sklearn.metrics
flask_cors